Footer Center and JavaScript Basic
the footer was centered and looking good so I applied that same code to every page in the site and now it looks awkwardly off-center but not exactly where it was before. Don't know if this is just glitching but I thought we could come back to it. And I got the Javascript collapsible menu to work for mobile!